NLP-main/requirements.txt
fastapi
uvicorn
spacy
docx2txt
pytesseract
pillow
cassandra-driver
pdf2image
llama-index
gradient
